LPS-106397 Rename expressions to expression.
