+++ modules/core/petra/petra-sql-dsl/src/main/java/com/liferay/petra/sql/dsl/Column.java
@@ -17,9 +17,9 @@ package com.liferay.petra.sql.dsl;
 import com.liferay.petra.lang.HashUtil;
 import com.liferay.petra.sql.dsl.ast.ASTNodeListener;
 import com.liferay.petra.sql.dsl.ast.impl.BaseASTNode;
-import com.liferay.petra.sql.dsl.expressions.Alias;
-import com.liferay.petra.sql.dsl.expressions.Expression;
-import com.liferay.petra.sql.dsl.expressions.impl.AliasImpl;
+import com.liferay.petra.sql.dsl.expression.Alias;
+import com.liferay.petra.sql.dsl.expression.Expression;
+import com.liferay.petra.sql.dsl.expression.impl.AliasImpl;
 
 import java.util.Objects;
 import java.util.function.Consumer;
