+++ modules/core/petra/petra-sql-dsl/src/main/java/com/liferay/petra/sql/dsl/ast/impl/DefaultASTNodeListener.java
@@ -17,8 +17,8 @@ package com.liferay.petra.sql.dsl.ast.impl;
 import com.liferay.petra.sql.dsl.Table;
 import com.liferay.petra.sql.dsl.ast.ASTNode;
 import com.liferay.petra.sql.dsl.ast.ASTNodeListener;
-import com.liferay.petra.sql.dsl.expressions.impl.Scalar;
-import com.liferay.petra.sql.dsl.expressions.impl.ScalarList;
+import com.liferay.petra.sql.dsl.expression.impl.Scalar;
+import com.liferay.petra.sql.dsl.expression.impl.ScalarList;
 import com.liferay.petra.sql.dsl.query.impl.Limit;
 
 import java.util.ArrayList;
