+++ modules/core/petra/petra-sql-dsl/src/main/java/com/liferay/petra/sql/dsl/expression/ElseEndStep.java
@@ -12,15 +12,23 @@
  * details.
  */
 
-package com.liferay.petra.sql.dsl.expressions;
+package com.liferay.petra.sql.dsl.expression;
+
+import com.liferay.petra.sql.dsl.ast.ASTNode;
+import com.liferay.petra.sql.dsl.expression.impl.ElseEnd;
+import com.liferay.petra.sql.dsl.expression.impl.Scalar;
 
 /**
  * @author Preston Crary
  */
-public interface Alias<T> extends Expression<T> {
+public interface ElseEndStep<T> extends ASTNode {
 
-	public Expression<T> getExpression();
+	public default Expression<T> elseEnd(Expression<T> expression) {
+		return new ElseEnd<>(this, expression);
+	}
 
-	public String getName();
+	public default Expression<T> elseEnd(T value) {
+		return elseEnd(new Scalar<>(value));
+	}
 
 }
