+++ modules/core/petra/petra-sql-dsl/src/main/java/com/liferay/petra/sql/dsl/expression/impl/DSLFunctionType.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.petra.sql.dsl.expressions.impl;
+package com.liferay.petra.sql.dsl.expression.impl;
 
 import java.util.Objects;
