+++ modules/core/petra/petra-sql-dsl/src/main/resources/com/liferay/petra/sql/dsl/expression/impl/packageinfo
