+++ modules/core/petra/petra-sql-dsl/src/test/java/com/liferay/petra/sql/dsl/ast/SQLDSLTest.java
@@ -20,22 +20,22 @@ import com.liferay.petra.sql.dsl.DSLQueryUtil;
 import com.liferay.petra.sql.dsl.Table;
 import com.liferay.petra.sql.dsl.ast.impl.BaseASTNode;
 import com.liferay.petra.sql.dsl.ast.impl.DefaultASTNodeListener;
-import com.liferay.petra.sql.dsl.expressions.Alias;
-import com.liferay.petra.sql.dsl.expressions.Expression;
-import com.liferay.petra.sql.dsl.expressions.Predicate;
-import com.liferay.petra.sql.dsl.expressions.WhenThenStep;
-import com.liferay.petra.sql.dsl.expressions.impl.AggregateExpression;
-import com.liferay.petra.sql.dsl.expressions.impl.AliasImpl;
-import com.liferay.petra.sql.dsl.expressions.impl.CaseWhenThen;
-import com.liferay.petra.sql.dsl.expressions.impl.DSLFunction;
-import com.liferay.petra.sql.dsl.expressions.impl.DSLFunctionType;
-import com.liferay.petra.sql.dsl.expressions.impl.ElseEnd;
-import com.liferay.petra.sql.dsl.expressions.impl.NullExpression;
-import com.liferay.petra.sql.dsl.expressions.impl.Operand;
-import com.liferay.petra.sql.dsl.expressions.impl.PredicateImpl;
-import com.liferay.petra.sql.dsl.expressions.impl.Scalar;
-import com.liferay.petra.sql.dsl.expressions.impl.ScalarList;
-import com.liferay.petra.sql.dsl.expressions.impl.WhenThen;
+import com.liferay.petra.sql.dsl.expression.Alias;
+import com.liferay.petra.sql.dsl.expression.Expression;
+import com.liferay.petra.sql.dsl.expression.Predicate;
+import com.liferay.petra.sql.dsl.expression.WhenThenStep;
+import com.liferay.petra.sql.dsl.expression.impl.AggregateExpression;
+import com.liferay.petra.sql.dsl.expression.impl.AliasImpl;
+import com.liferay.petra.sql.dsl.expression.impl.CaseWhenThen;
+import com.liferay.petra.sql.dsl.expression.impl.DSLFunction;
+import com.liferay.petra.sql.dsl.expression.impl.DSLFunctionType;
+import com.liferay.petra.sql.dsl.expression.impl.ElseEnd;
+import com.liferay.petra.sql.dsl.expression.impl.NullExpression;
+import com.liferay.petra.sql.dsl.expression.impl.Operand;
+import com.liferay.petra.sql.dsl.expression.impl.PredicateImpl;
+import com.liferay.petra.sql.dsl.expression.impl.Scalar;
+import com.liferay.petra.sql.dsl.expression.impl.ScalarList;
+import com.liferay.petra.sql.dsl.expression.impl.WhenThen;
 import com.liferay.petra.sql.dsl.query.DSLQuery;
 import com.liferay.petra.sql.dsl.query.FromStep;
 import com.liferay.petra.sql.dsl.query.GroupByStep;
